Remove free_vars from core and move fresh_var
