+++ lang/core_lang/src/syntax/names.rs
@@ -1,8 +1,22 @@
 use crate::traits::*;
 
+use std::collections::HashSet;
+
 pub type Var = String;
 pub type Covar = String;
 pub type Name = String;
+
+#[must_use]
+pub fn fresh_var(used_vars: &mut HashSet<Var>, base_name: &str) -> Var {
+    let mut n = 0;
+    let mut new_var: Var = format!("{base_name}{n}");
+    while used_vars.contains(&new_var) {
+        n += 1;
+        new_var = format!("{base_name}{n}");
+    }
+    used_vars.insert(new_var.clone());
+    new_var
+}
 
 impl SubstVar for Var {
     type Target = Var;
