+++ lang/core_lang/src/traits/mod.rs
@@ -1,11 +1,9 @@
 mod focus;
-mod free_vars;
 mod substitution;
 mod typed;
 mod uniquify;
 
 pub use focus::{bind_many, Bind, Continuation, Focusing, FocusingState};
-pub use free_vars::{fresh_var, FreeV};
 pub use substitution::{Subst, SubstVar};
 pub use typed::Typed;
 pub use uniquify::Uniquify;
