+++ lang/fun2core/src/terms/constructor.rs
@@ -6,7 +6,7 @@ use crate::{
 };
 use core_lang::syntax::{
     term::{Cns, Prd},
-    types::Ty,
+    Ty,
 };
 
 impl CompileWithCont for fun::syntax::terms::Constructor {
