+++ lang/fun2core/src/terms/label.rs
@@ -6,7 +6,7 @@ use crate::{
 };
 use core_lang::syntax::{
     term::{Cns, Prd},
-    types::Ty,
+    Ty,
 };
 
 impl CompileWithCont for fun::syntax::terms::Label {
