+++ lang/fun2core/src/terms/lit.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 use crate::definition::CompileWithCont;
 use core_lang::syntax::{
     term::{Cns, Prd},
-    types::Ty,
+    Ty,
 };
 
 impl CompileWithCont for fun::syntax::terms::Lit {
